+++ src/ani/index.ts
@@ -1,2 +1,1 @@
-export type { AniFrame } from "./AniFile";
-export { AniFile } from "./AniFile";
+export { generateTresFiles } from "./tres";